Refactor: Overhaul SMS status receiver for reliability and security This commit refactors the mechanism for handling SMS status updates (sent, delivered, failed) to fix a critical reliability bug and improve security.
The previous implementation created a new BroadcastReceiver for each outgoing SMS. This pattern was unreliable under heavy load, causing some messages to get stuck indefinitely in the "sending" state.

The new implementation addresses this by:
1.  Introducing a single, centralized `SmsStatusReceiver` to handle all status updates, eliminating race conditions.
2.  Dynamically registering this receiver within the `GatewayService` lifecycle instead of statically in the `AndroidManifest.xml`. This closes a security vulnerability by ensuring the receiver is not exported to other applications.
3.  Using `ContextCompat.registerReceiver` to manage the registration, which simplifies the code and ensures compatibility with the latest Android version requirements (Android U/14+).

This change makes the SMS sending process significantly more robust and secure.
