+++ app/src/androidTest/java/com/quansoft/smsgateway/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinsmsgateway
+package com.quansoft.smsgateway
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.kotlinsmsgateway", appContext.packageName)
+        assertEquals("com.quansoft.kotlinsmsgateway", appContext.packageName)
     }
 }
