+++ app/src/main/java/com/quansoft/smsgateway/data/SmsDao.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinsmsgateway.data
+package com.quansoft.smsgateway.data
 
 import androidx.room.Dao
 import androidx.room.Insert
