+++ app/src/main/java/com/quansoft/smsgateway/service/BootReceiver.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinsmsgateway.service
+package com.quansoft.smsgateway.service
 
 import android.content.BroadcastReceiver
 import android.content.Context
