+++ app/src/main/java/com/quansoft/smsgateway/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.example.kotlinsmsgateway.ui.theme
+package com.quansoft.smsgateway.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
