+++ app/src/test/java/com/quansoft/smsgateway/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinsmsgateway
+package com.quansoft.smsgateway
 
 import org.junit.Test
